% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRrpart.R
\name{CRrpart}
\alias{CRrpart}
\title{Recursive partitioning with hypothesis testing at each split for a competing risks endpoint}
\usage{
CRrpart(
  x,
  ftime,
  fstatus,
  sig.level = 0.05,
  n.splits = 500,
  minbucket = NULL,
  support = NULL,
  failcode = 1,
  cencode = 0,
  p.adj = FALSE
)
}
\arguments{
\item{x}{Predictor variables}

\item{ftime}{Failure/censoring time variable}

\item{fstatus}{Failure/censoring event variable}

\item{sig.level}{Significance level (default = 0.05)}

\item{n.splits}{Maximum number of splits (default =500)}

\item{minbucket}{Minimum number of observations needed in each daughter node}

\item{support}{Percent of observations needed in each daughter node}

\item{failcode}{Code of fstatus corresponding to the event of interest (default = 1)}

\item{cencode}{Code of fstatus corresponding to censored observations (default = 0)}

\item{p.adj}{Adjust p-value for multiple testing? (default =FALSE). The adjustment is based upon the the distribution of the maximum log-likelihood statistic. It adjusts for the number of input variables, not the number of valid data splits.}
}
\value{
def: Terms used to create the recursive partitions

pvalue: pvalues for the significant terms

call: Input function call parameters and data

class.id: Subgroup identifier
}
\description{
Recursively partition a dataset with competing risks endpoint using the maximum pseudo log-likelihood among potential splits to determine optimal split. The partitioning will stop when there are not more statistically significant splits. The package uses the competing risk analysis formulation and functions from the 'cmprsk' package.
}
\examples{
\dontrun{
library(CRrpart)
g=CRrpart(minbucket=NULL,support=.2,ftime=bmtcrr$ftime,fstatus=bmtcrr$Status,
cencode=0,failcode=1,x=bmtcrr[,c('Sex','D','Phase','Age','Source')],
n.splits=500, sig.level=0.05,p.adj=FALSE)
}
}
\seealso{
\link[cmprsk]{crr}, \link[cmprsk]{cuminc}
}
