% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRrpart.R
\name{plot.CRrpart}
\alias{plot.CRrpart}
\title{Plot the cumulative incidence by output classes}
\usage{
\method{plot}{CRrpart}(x, failcode = 1, ...)
}
\arguments{
\item{x}{CRrpart object}

\item{failcode}{Code of fstatus corresponding to the event of interest (default = 1)}

\item{...}{arguments to be passed to methods}
}
\description{
Predict output classes to a new database
}
\examples{
\dontrun{
library(CRrpart)
g=CRrpart(minbucket=NULL,support=.2,ftime=bmtcrr$ftime,fstatus=bmtcrr$Status,
cencode=0,failcode=1,x=bmtcrr[,c('Sex','D','Phase','Age','Source')],
n.splits=500,sig.level=0.05,p.adj=FALSE)
plot(object=g)
}
}
