% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRrpart.R
\name{predict.CRrpart}
\alias{predict.CRrpart}
\title{Predict output classes to a new database for CRrpart object}
\usage{
\method{predict}{CRrpart}(object, newdata = NULL)
}
\arguments{
\item{object}{CRrpart object}

\item{newdata}{new data with same input predictors as the CRrpart object}
}
\value{
outclass: output class for observation
}
\description{
Predict output classes to a new database
}
\examples{
\dontrun{
library(CRrpart)
g=CRrpart(minbucket=NULL,support=.2,ftime=bmtcrr$ftime,fstatus=bmtcrr$Status,
cencode=0,failcode=1,x=bmtcrr[,c('Sex','D','Phase','Age','Source')],
n.splits=500,sig.level=0.05,p.adj=FALSE)
predict(object=g)
}
}
