% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRrpart.R
\name{print.CRrpart}
\alias{print.CRrpart}
\title{Print an CRrpart object}
\usage{
\method{print}{CRrpart}(x, tim = 24, failcode = object$call$failcode, ...)
}
\arguments{
\item{x}{CRrpart object}

\item{tim}{The time at which to report cumulative incidence of event failcode}

\item{failcode}{Code of fstatus corresponding to the event of interest (default = 1)}

\item{...}{arguments to be passed to methods}
}
\description{
Print an CRrpart object
}
\examples{
\dontrun{
library(CRrpart)
g=CRrpart(minbucket=NULL,support=.2,ftime=bmtcrr$ftime,fstatus=bmtcrr$Status,
cencode=0,failcode=1,x=bmtcrr[,c('Sex','D','Phase','Age','Source')],
n.splits=500, sig.level=0.05,iter=2000,p.adj=FALSE)
}
}
